Fixes, Volatile Objects, and Mementos
Fixes
Fixed any old station logic being uncompileable with the new Station namespace.
Volatile Objects
Volatile Objects are the pure C# objects we will be using in the threaded environment to protect Unity's thread from locking on accessing it's monobehaviours, which these will be added to.
Memento
Mementos are the new property standard for any object that needs to have previous states of said object stored, for statistical analysis, or state saving.