VolatileModule Distribution and Sequencing
I am starting to flesh out the functional elements of how this whole system is going to work, I have a small issue with how I am going to get the .Update() to get a Request since it will be executed by a thread, from a void delegate(void). Not entirely sure what the solution is there, yet, might be very obvious.

Distribution, so I wrote some functional, but incredibly distasteful code. I am going to change the code in the coming push so it functions by properly accounting for hardpoint filter volume transfer limits, as implicit in the volume amounts defined in the resourcestack[] Filters. The current system updates the module with the most optimistic inventory circumstances, then when the recieving modules update and attempt to recieve the resources, they have to throw all the unaccepted stuff back in another UpdateModuleRequest, which I have yet to implement, as the new system wont need it, and I really don't feel comfortable having to use this gross solution..