Prepare and Events
Prepare is the events system, and as such I felt it fitting to make it its own 'package'.
Event is a request, but of a specific class, with a unique logic of operation (Do must propegate the event to connected hardpoints)

Up next, Memento thread safety, and refactoring locking to lock on smallest denominator instead of entire object.