VolatileHardpoint, Memento, VolatileModule, and UpdateModuleRequest
VolatileHardpoint
Updated to properly utilise the fixed IInventory (changes incorrect check against null, and removed array accessor for GetResource()).

Memento
Change() - Change allows you to change the last state of the Memento.
Implemented because often in the update cycle of modules do multiple modules change values on one module each update, I would rather see net changes on the module than changes per request.

VolatileModule
With new Change() on Memento I no longer need the secretive mementos with standard float properties.
Did this because to implement the desired changes to module mementos I needed access to the mementos proper, not the type cast accessor to them.

UpdateModuleRequest
Changed to reflect the new memento standards.
This means that the module first on update sets its power this update to 0
then all the requests change that value to whatever it is plus their energyIn, this means that when they are all done that, I get one comprehensive state, which tells me when it was done processing energyIn requests, and how much power in total was received this frame. Easy peasy.