Globals and CommandModules
I have made a probably temporary solution for a place to look to for global values, that aren't constant. Such as SimulationSpeed, a memento which tracks what speed the game is set to simulate at. Necessary for command module updates.

CommandModule
Queues a thread Update of the VolatileModule threaded (which in this case will actually be a VolatileCommandModule).

VolatileCommandModule
Implements the abstract OverridableUpdate() giving it updateSequence incrementation logic (which causes the cascaded updates).

VolatileModule
Added missing null check for connections on hardpoints
Updated incorrect comments (removed in this case).