SI Adherence and VolatileModule
Adding a solution for correctly using SI units, and offering an easy to use solution for output to string, as well as hopefully, math.

VolatileModule
Added the missing all module must have parameters, as defined in the GDD.

Statistic
Statistic is a memento with a maximum value, used for any circumstance where before I had a maximum value and a memento, to encapsulate that repeating logic in a simple class.

UpdateModuleRequest
Started implementing full distribution of resources, but stopped to implement an SI unit solution.

SIUnit
Working to cover the bases of all necessary SI units:
Gram
Kelvin
Metre
Mole
Watt
The idea here is that if we say don't need Mole, we implement the one we do, and make sure it can math nicely with other SI units. In my case, Watt, as the actual measure in SI is ampere, which goes to joule, to watt, but as I never plan on implementing anything below watthours, I will skip em.