SIUnit and Friends
Continuing to implement the SIUnit and a library of fellow SI units of measurement, like Gram, Second, Hour, Kelvin, Metre, Mole, Watt, and WattHour (The first of the 'compound' SIUnits).

Implementation looks primarily like adding extensive operator overloads to the SIUnits to make them compatible for math in specific ways, and for conversion between compatible units. Some tricky parts currently would be the WattHour multiply, which I am not sure I will ever need. Also it has a watt element and a hour element, which means it can actually have say 356W7h which is meaningless, so I should probably think on that, and get rid of it maybe.