Systems
I started implementing some of the systems.
Pressurisation -> Handles pressurising volumes with gasses.
Electrical Grid -> Handles energy distribution between connected systems.
Connections -> Handles connections between systems.
Atmospherics -> Handles atmospheric variables (Pressure and Temperature in a module) and interfacing with the appropriate systems to enforce them.

Pushing because after a length debate with myself, and a sleep, I have decided that I implemented it wrong. It always bothered me that the states handled updating their dirtied values, while it was a neat way of allowing states to be passed around, changed, and have the changes reflect in the source of the state, I have decided it was wrong, states are states.
So I will begin removing any logic from the states, and moving it to the system interfaces (IPressurisation, etc).