SIUnit switched to struct
That was a lot of work.
So I had to move the explicit float cast, and the implicit string cast to declarations on every SIUnit, as C# structs can have no inheritance.
Each SIUnit is now inheriting from ISIUnit which guarantees that a Value can be accessed through Value() and Value(float f).
SIUnit now has static helpers to ensure the previously inherited functions are still only defined once.