Src Removal, Struct SIUnits and States
Switched standardised method calls previously coming from inheritance over to interface implementation in both States and SIUnits, to ensure they are as fast as possible for value types (this has caused a small issue, we can no longer make them const, a limitation of struct implementations in C#).

Removed Src as a namespace from everything save resources (for git compatibility with Mark)

Added more operator overloads to reduce the number of errors in the project, with math that no longer works since making float casting explicit on all SIUnits.

Switched state references to anything to a GUID, added a GUID extension that can get a state for any GUID if that GUID has a state of the specified type.

Implemented IState for any generic state methodologies.

Improved System, it now acts as the source of all Systems state storage, they will each register themselves here, and anyone will be able to go here and query specific states for specific systems.

Next up to implement the handlers for Systems, and the state registration process. As well to finish removing all SIUnit errors so testbed scenes can be constructed in Unity, and live testing of threading and systems can begin.