Systems - Pressurisation
Further refining of pressurisation.
Having done more extensive research of Ideal Gas Law vs Van der Waal equation of states, I have determined that so long as atmospheric pressurisation systems are limited within certain gas, pressure, and temperature limitations I can avoid most of the problems inherent in the Ideal Gas Law, this however will make atmospheric readings approximative, and not entirely accurate (quite possibly).

After doing some math to test my implementation of the Ideal Gas Law, and having values that I could not verify returned, I realised that to correctly implement it I must stick more strictly to the units expected, that is I needed some way of converting Metre3 to Litre and back, I have implemented a very simple conversion class Litre to handle this, such that while all equations continue to be handled with Metre3 and the suite of operator overloads that are prepared for it, values enter the Pressurisation system in Litres, which are measured as 10x10x10 cubic centimeters. Ask me for my validation of Ideal gas law document for proof of this error, and subsequent solution.

Fixed GuidExtensions to support the new Systems.
Added Litre to both Name and Symbol of Literals.
Tidied up PressurisationState.

Now feel confident to offer a rundown of what Pressurisation does.

Pressurisation
Pressurisation handles the pressurisation specific state of any object that is being handled by Pressurisation, by attached Guid, through Systems.
Pressurisation is capable of, given an Atmospherics system, and an Inventory, determine the pressure inside a volume at a given temperature, as well it can determine the number of moles of gas present (This does not account for non-ideal gasses, those will have to be handled on a case specific bases 02 instead of just 0, I assume a new resource with twice the molar mass will solve for this).
It can also determine the amount of resources necessary to achieve the desired pressure temperature based on the current resources in the volume.
Essentially it can calculate anything pressurisation dependant, and as such has a subsequent state that only stores those values, as well as references to the other systems necessary for complete calculations.
It was my executive decision to reduce value copying by removing any duplicate values other systems states already store, and properly manage, such as pressure desired (atmospherics), temperature (atmospherics), and volume (inventory).